Bug 969865 - Fix the allocation handling of frame time recording
FramesTimingRecording::mIntervals should use fallible allocations, as
the code which is using it is prepared to handle OOM conditions.
However, this handling is not required for the array argument passed to
LayerManager::StopFrameTimeRecording, since that array is coming from
the IPDL protocol, and as of bug 967167 we handle the OOM conditions at
the IPDL protocol layer properly, so no extra OOM handling is necessary
here.

https://bugzilla.mozilla.org/show_bug.cgi?id=969865