Bug 1508759 - Add support for hooking up permission prompts that don't use SitePermissions to TemporaryPermissions; r=johannh
Normally, permission prompts would define a permissionKey attribute in order
to get integrated with SitePermissions.  Since SitePermissions is internally
hooked up to TemporaryPermissions, such permission prompts do not need any
extra handling for taking benefit from the temporary permissions infrastructure.

For the Storage Access API, however, we're not going to use SitePermissions,
and instead Gecko will be in charge of defining the required permissions in the
permission manager database when the prompt is responded to with an Allow
action.  This means that by default we won't be integrated with the temporary
permissions setup either.

This patch allows prompts to define a new way to get integrated with temporary
permissions.  That is, through defining a new permissionDynamicKey attribute.
This attribute will return a dynamically generated key that is unique to each
instance of the prompt.  TemporaryPermissions will do the expected work in
order to ensure that each unique prompt key will be automatically blocked if
a previous instance of the same prompt type has been blocked in the current
tab.

Note that this doesn't yet include support for showGloballyBlocked or
permitTemporaryAllow since those features aren't needed for our use case.

Differential Revision: https://phabricator.services.mozilla.com/D12466