Bug 1540810 - Part 2: Add a ContentBlockingNotifier API for notifying a document that third-party content is now partitioned from its HTTP channel, r=baku,mayhemer
This API would allow us to notify a generic channel object about the
fact that the storage for third-party content should be partitioned
dynamically from now on, and communicate this with the document the
channel belongs to, irrespective of whether the notification is emitted
from the parent or the content process.

This will ensure that future channels created from the browsing context
to which this document belongs will honour the partitioned storage
semantics.

Differential Revision: https://phabricator.services.mozilla.com/D54573