Bug 1540810 - Part 5: Partition 3rd party storage when the canonical host name heuristic was used; r=baku
This patch makes the heuristic added in the previous part safe to
deploy, by ensuring that the storage access granted is always a
partitioned storage grant, tied to the top-level document's eTLD+1.

This uses the dynamic FPI infrastructure to dynamically turn on FPI only
for the third-party subresources that the heuristic is being applied to.

Differential Revision: https://phabricator.services.mozilla.com/D54576