Bug 1540810 - Part 3: Add a Necko API for retrieving the canonical host name of channels, and keeping tracking of the top-level channel's canonical host name; r=mayhemer,baku
These Necko APIs can be used for keeping track of the top-level
document's canonical host name (CNAME), as queried from the DNS
subsystem.  Additionally it will provide access to the canonical host
name of the channel URI, which will be used in future anti-tracking
interventions (see bug 1592727 for example).

For the heuristic implemented in this bug, we only really need the
top-level document's canonical host name information.

Differential Revision: https://phabricator.services.mozilla.com/D54574