Bug 1540810 - Part 4: Promote third-party sub-resouces from domains that are CNAME'd to be same-eTLD+1 as the top-level document's origin to be considered as first-party sub-resources; r=baku
This is the implementation of the main heuristic that fixes the breakage
reported in this bug.  With parts 1-4 applied, we support successful
login to ZenDesk support forums.

Note that this heuristic is landing disabled by default on late
beta/release builds, and will be turned on after further testing in a
future bug.

Differential Revision: https://phabricator.services.mozilla.com/D54575