Allow calling WhenCanonicalHostNameAvailable() before a channel has been opened
This is done by kicking off DNS prefetch when that function is called,
and preparing to start proxy resolution and connecting to the network if
needed when the response to the DNS query becomes available.

This will allow the URL classifier to use the results of the DNS query
as input, such as the canoniocal host name.